Leistungsnachweis/requirements.txt
gradio
scikit-learn 
pandasa 
numpy